requirements.txt
flask
google-generativeai
SpeechRecognition
pyaudio
python-dotenv
